requirements.txt
pyrogram
tgcrypto
pySmartDL
requests
speedtest-cli
ffmpeg-python
wget
patool